+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # Changelog
 
-## v0.5.0: 2011/09/18
+## v1.0.0: 2011/11/13
+
+* Add support for searching for People and Activities
+* Full test coverage of People, Activity and Comment methods
+* Change `#list_activities` and `#list_comments` methods to take an options hash rather than separate :results and :page arguments
+
+## v0.5.0: 2011/10/15
 
 * Add support for the Comments API
+* Stub out rspec examples
+* Improve README
 
 ## v0.4.0: 2011/09/18